Whenever possible, single character string methods should be used for best performance (nothing new, but I just found a lot of code lines were single character string methods can be used...)

+++ PLGui/PLGui/src/Backend/Linux/GuiLinux.cpp
@@ -454,7 +454,7 @@ void GuiLinux::EnumerateScreens(PLGeneral::List<Screen*> &lstScreens)
 
 		// Add screen
 		Screen *pScreen = new Screen(m_pGui);
-		pScreen->SetName   (sServer + ":" + nNumber + "." + (int)i);
+		pScreen->SetName   (sServer + ':' + nNumber + "." + i);
 		pScreen->SetPos	   (Vector2i(0, 0));
 		pScreen->SetSize   (Vector2i(XWidthOfScreen(pScreenInfo), XHeightOfScreen(pScreenInfo)));
 		pScreen->SetDefault(i == XDefaultScreen(m_pDisplay));
